Analysis_Code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.23.0=pypi_0
aiofiles=23.2.1=pypi_0
altair=5.3.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
blis=0.7.11=pypi_0
bokeh=3.1.1=pypi_0
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
catalogue=2.0.10=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpathlib=0.16.0=pypi_0
cloudpickle=3.0.0=pypi_0
colorama=0.4.6=pypi_0
colorcet=3.1.0=pypi_0
confection=0.1.4=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.12.1=pypi_0
cymem=2.0.8=pypi_0
daam=0.2.0=pypi_0
dask=2023.5.0=pypi_0
datashader=0.15.2=pypi_0
datashape=0.5.2=pypi_0
diffusers=0.21.2=pypi_0
einops=0.7.0=pypi_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.110.1=pypi_0
ffmpeg=4.3=hf484d3e_0
ffmpy=0.3.2=pypi_0
filelock=3.13.4=pypi_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
gradio=4.26.0=pypi_0
gradio-client=0.15.1=pypi_0
h11=0.14.0=pypi_0
holoviews=1.17.1=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.22.2=pypi_0
idna=3.7=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.4.0=pypi_0
inflect=7.2.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=pypi_0
joblib=1.4.0=pypi_0
jpeg=9e=h5eee18b_1
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
linkify-it-py=2.0.3=pypi_0
llvmlite=0.41.1=pypi_0
locket=1.0.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
mdit-py-plugins=0.4.0=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
more-itertools=10.2.0=pypi_0
multipledispatch=1.0.0=pypi_0
murmurhash=1.0.10=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
nltk=3.8.1=pypi_0
numba=0.58.1=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_0
orjson=3.10.1=pypi_0
outcome=1.3.0.post0=pypi_0
packaging=24.0=pypi_0
pandas=2.0.3=pypi_0
panel=1.2.3=pypi_0
param=2.1.0=pypi_0
partd=1.4.1=pypi_0
pillow=9.5.0=pypi_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
preshed=3.0.9=pypi_0
psutil=5.9.8=pypi_0
pyct=0.5.0=pypi_0
pydantic=2.7.0=pypi_0
pydantic-core=2.18.1=pypi_0
pydub=0.25.1=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.9=pypi_0
pytorch=1.10.2=py3.8_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyviz-comms=3.0.2=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.34.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.0=pypi_0
ruff=0.3.7=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
selenium=4.19.0=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=68.2.2=py38h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pypi_0
smart-open=6.4.0=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
spacy=3.7.4=pypi_0
spacy-legacy=3.0.12=pypi_0
spacy-loggers=1.0.5=pypi_0
sqlite=3.41.2=h5eee18b_0
srsly=2.4.8=pypi_0
starlette=0.37.2=pypi_0
tbb=2021.8.0=hdb19cb5_0
thinc=8.2.3=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=pypi_0
tomlkit=0.12.0=pypi_0
toolz=0.12.1=pypi_0
torchvision=0.11.3=py38_cu113
tornado=6.4=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.30.2=pypi_0
trio=0.25.0=pypi_0
trio-websocket=0.11.1=pypi_0
typeguard=4.2.1=pypi_0
typer=0.9.4=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
uc-micro-py=1.0.3=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
wasabi=1.1.2=pypi_0
wcwidth=0.2.13=pypi_0
weasel=0.3.4=pypi_0
webencodings=0.5.1=pypi_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py38h06a4308_0
wsproto=1.2.0=pypi_0
xarray=2023.1.0=pypi_0
xyzservices=2024.4.0=pypi_0
xz=5.4.6=h5eee18b_0
zipp=3.18.1=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
